Increase tmux buffer size to 5000
This is untested in different versions of tmux. I have not yet looked
into which version of tmux this feature was added in. If it is a feature
that has been around since tmux 1.8 I will keep it where it is. Currently
the oldest version of tmux that I have to support is 1.8.

There is also the possibility that this will cause too much RAM to be
used (I have a tendency to use lots of tmux sessions at once). This will
need to be monitored as well.