Correct spell file error and python spell check
This corrects vim to point to the correct spelling files. It fixes a bug
that was preventing them from being evaluated. It also adds the ability
for vim to check for spelling errors in Python.

This will add 3 spell files to the spell file path:

1. If Git exists on the system and we are in a repo, set the first file
   to `.local.en.utf-8.add` in the root of the repo (a repo-wide spell
   file). If Git is not installed or we aren't in a repo, set the spell
   file adjacent to the current file.
2. A file specific to the particular language that we are using (i.e.,
   there will be one for C++, Python, XML, etc.). These files will be
   placed in the location
   `$XDG_CONFIG_HOME/nvim/spell/filetype/en.utf-8.add`
3. A file that is global - this is used for words that are common
   across our work realm that may or may not be used elsewhere (for
   example - acronyms). This file will be placed in the location
   `$XDG_CONFIG_HOME/nvim/spell/en.utf-8.add`

The name of the file matters because Vim uses the file name to determine
what language (en = English), what encoding (utf-8=UTF-8), and purpose
(add=additional).