Add spaceship theme
Spaceship has a dependency on powerline to display fonts correctly. On
RHEL 8+, powerline is an rpm provided by EPEL that can be installed,
which is what we check for. On older versions of RHEL (7.9 or older),
powerline is not an rpm and would require too much effort to support
(especially on soon to be EOL OS's). As for MacOS, there is a gitlab
repository that can be cloned into fonts/ that will be source
automagically. Because we ignore anything that we don't explicitly add
in the repository, it will never show up as a diff and can easily be
cloned.

For MacOS:

cd /path/to/configs
mkdir fonts && cd fonts
git clone https://github.com/powerline/fonts.git